% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_tools.R
\name{ci.FI}
\alias{ci.FI}
\title{Fisher information confidence intervals}
\usage{
ci.FI(model)
}
\arguments{
\item{model}{An mle2 model fit}
}
\value{
A 95 percent confidence interval
}
\description{
Calculate 95 percent confidence intervals for model parameters, based
on Fisher Information.
}
