% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_tools.R
\name{confint.FI}
\alias{confint.FI}
\title{Fisher information confidence intervals}
\usage{
\method{confint}{FI}(model)
}
\arguments{
\item{model}{An mle2 model fit}
}
\value{
A 95% confidence interval
}
\description{
Calculate 95% confidence intervals for model parameters, based
on Fisher Information.
}

