% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_mle.R
\name{convergeQ}
\alias{convergeQ}
\title{Convergence test}
\usage{
convergeQ(x)
}
\arguments{
\item{x}{An mle2 mode object}
}
\value{
The convergence status of the model (1 or 0)
}
\description{
Check whether a particular mle2 model converged during fitting
}

