% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_distributions_and_links.R
\name{dbeta2}
\alias{dbeta2}
\title{Mean-parameterized beta distribution}
\usage{
dbeta2(x, mu, phi, log = FALSE)
}
\arguments{
\item{x}{Calculate the probability density at these values}

\item{mu}{The mean of the beta distribution}

\item{phi}{Dispersion parameter}

\item{log}{logical; return the ln of the probability density?}
}
\description{
Mean-parameterized beta distribution
}

