% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_distributions_and_links.R
\name{dgamma2}
\alias{dgamma2}
\title{Mean-parameterized gamma distribution}
\usage{
dgamma2(x, mu, s, log = F)
}
\arguments{
\item{x}{Calculate the probability density at these values}

\item{mu}{The mean of the gamma distribution}

\item{s}{Dispersion parameter}

\item{log}{logical; return the ln of the probability density?}
}
\description{
Note: parmeter s is not the variance of the distribution. 
Variance = mu + mu^2/s
}

