% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_distributions_and_links.R
\name{dzabeta}
\alias{dzabeta}
\title{Zero-altered beta distribution
	
Code for vectorized zero altered beta distribution that avoids
errors in dbeta, and can handle log values.  Follows mean and
precision parameterization given in Ramalho et al. 2009}
\usage{
dzabeta(x, z, mu, phi, log = FALSE)
}
\arguments{
\item{x}{}

\item{log}{}
}
\description{
Zero-altered beta distribution
	
Code for vectorized zero altered beta distribution that avoids
errors in dbeta, and can handle log values.  Follows mean and
precision parameterization given in Ramalho et al. 2009
}

