% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_tools.R
\name{gen.R2}
\alias{gen.R2}
\title{Generalized R2 estimation, based on Cox & Snell 1989}
\usage{
gen.R2(m1, m2)
}
\arguments{
\item{m1}{An intercept-only model}

\item{m2}{A more complex model}
}
\value{
Generalized R2 value
}
\description{
Comparisons across models invoking different error distributions
is not recommended at present.
}

