% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_mle.R
\name{grid.mle2}
\alias{grid.mle2}
\title{Maximum likelihood model fitting using a grid of starting guesses}
\usage{
grid.mle2(minuslogl, grids, start, data, ...)
}
\arguments{
\item{minuslogl}{A negative log likelihood function; passed to mle2}

\item{grids}{A list of multiple starting guesses for parameters found in minuslogl}

\item{start}{A list of single starting guesses for parameters in minuslogl}

\item{data}{A data frame containing the data used in the minuslogl function}
}
\value{
res.mat A matrix of coefficient estimates, model AIC values, and convergence status

res.model An indexed list of mle2 objects, one for each fit, ordered as in res.mat

res.best The best model selected from list of res.model
}
\description{
Designed to perform mle2 fits for a grid of initial parameter
guesses, to aid situations where good starting values are either
difficult to obtain, or when a large number of fits need to be
conducted in an automated fasion. A large number of fits are 
performed; from the sbuset of resulting models that successfully
converge, the best model is selecte based on AIC comparison.
}
\details{
This function is essentially a wrapper for mle2() from 
Ben Bolker's bbmle package.
}
\examples{
See examples in grid.mle2_wexamples.R

}

