% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleTools.R
\docType{package}
\name{mleTools}
\alias{mleTools}
\title{mleTools: A package for extending maximum likelihood analysis tools (in bbmle).}
\description{
The mleTools package does X, Y, and Z. (see: http://r-pkgs.had.co.nz/man.html#man-packages)
}
\section{grid.mle2 functions}{

The grid.mle2 functions does ...
}

